+++ config_default.json
@@ -1,0 +1,7 @@
+{
+    "hamqth": {
+        "username": "put username here",
+        "password": "put password here"
+    },
+    "oofs": "0"
+}
