+++ hamtheman.py
@@ -2,18 +2,44 @@
 # Purpose: It performs various ham radio-related tasks
 
 
+import json
 import discord
 
-import core
-from core import bot, bonk, boonk
 from commands import lookup, misc, morse
+
+
+'''
+STARTUP THINGS
+'''
+
+# master config
+config = {}
+with open('config.json', 'r') as f:
+    config = json.load(f)
+    print('config loaded')
+
+# for calculating uptime
+start_time = time.time()
+
+# for looking up callsigns
+call_lookup = htmlookup.HtmLookup()
+
+
+'''
+DISCORD STUFF
+'''
+
+
+
 
 
 # startup stuff
 @bot.event
 async def on_ready():
-    await bot.change_presence(activity=discord.Game(name="with Baofengs | htm help"))
+    await bot.change_presence(
+        activity=discord.Game(name="with Baofengs | htm help"))
     print('Shaking and Baking')
+
 
 # every message goes through here
 @bot.event
@@ -22,8 +48,9 @@ async def on_message(message):
     message.content = message.content.lower()
 
     # get the bonks, boonks, and the oofs
+    # TODO: Make a thread that periodically saves the oof count
     if message.content == 'oof':
-        core.oof_count += 1
+        config['oofs'] += 1
         await message.channel.send('rip')
     elif message.content == 'bonk':
         await message.channel.send(bonk)
@@ -39,3 +66,27 @@ with open('keys/discord.txt', 'r') as f:
     lines = f.readlines()
     if len(lines) == 1:
         bot.run(lines[0].strip())
+
+
+'''
+STRINGS AND STUFF
+'''
+
+
+bonk = (':regional_indicator_b: '
+        ':regional_indicator_o: '
+        ':regional_indicator_n: '
+        ':regional_indicator_k:')
+
+boonk = (':regional_indicator_b: '
+         ':regional_indicator_o: '
+         ':regional_indicator_o: '
+         ':regional_indicator_n: '
+         ':regional_indicator_k:     '
+         ':regional_indicator_g: '
+         ':regional_indicator_a: '
+         ':regional_indicator_n: '
+         ':regional_indicator_g:')
+
+bot = discord.ext.commands.Bot(command_prefix=config['prefix'])
+bot.remove_command('help')
